chat_bot_api/requirements.txt
apify-client
chromadb
langchain
langchain-community
langchain-core
langchain-openai
langchain_google_vertexai
langchain_google_community
langgraph
openai
python-dotenv
streamlit
tiktoken
fastapi
fastapi[standard]
uvicorn[standard]
typing
pydantic
vertexai
jq